third_party/blink/web_tests/accessibility/spinbutton-value-expected.txt

42

This tests that a spin button supports range value attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS endsWith(axSpin.intValue, '5') is true
PASS endsWith(axSpin.minValue, '1') is true
PASS endsWith(axSpin.maxValue, '9') is true
PASS endsWith(axSpin.valueDescription, '5 of 9') is true
PASS axSpin.name is axUntitled.name
PASS successfullyParsed is true

TEST COMPLETE
